+++ http/gorilla/mux/module.go
@@ -1,4 +1,4 @@
-package gorilla
+package fxmux
 
 import (
 	"github.com/gorilla/mux"